Remove systems from Sprocket
