+++ Anvil/anvil.cpp
@@ -9,7 +9,6 @@
 #include <Sprocket/Graphics/camera.h>
 #include <Sprocket/Graphics/material.h>
 #include <Sprocket/Scene/Loader.h>
-#include <Sprocket/Scene/Systems/input_system.h>
 #include <Sprocket/Scene/meta.h>
 #include <Sprocket/UI/ImGuiXtra.h>
 #include <Sprocket/Utility/FileBrowser.h>
@@ -218,7 +217,7 @@ void Anvil::on_render()
             if (ImGui::MenuItem("Run")) {
                 d_activeScene = std::make_shared<spkt::scene>();
 
-                spkt::input_system_init(d_activeScene->registry, d_window);
+                anvil::input_system_init(d_activeScene->registry, d_window);
                 spkt::copy_registry(d_scene->registry, d_activeScene->registry);
 
                 d_activeScene->systems = {
@@ -228,11 +227,11 @@ void Anvil::on_render()
                     anvil::animation_system,
                     anvil::delete_below_50_system,
                     anvil::clear_events_system,
-                    spkt::input_system_end
+                    anvil::input_system_end
                 };
 
                 d_activeScene->event_handlers = {
-                    spkt::input_system_on_event
+                    anvil::input_system_on_event
                 };
 
                 d_playingGame = true;
